Forms/Product/requirements.txt
Image==1.5.33
matplotlib==3.4.2
numpy==1.20.1
opencv_python==4.5.1.48
pandas==1.3.3
pdf2image==1.14.0
Pillow==8.4.0
pytesseract==0.3.7
rich==10.13.0
